refactor(card-register): 폴더명을 CardRegisterPage로 변경
