+++ src/feature/CardRegisterPage/components/CardPreview/CardExpiryDateDisplay/CardExpiryDateDisplay.stories.tsx
