+++ src/feature/CardRegisterPage/components/CardPreview/CardExpiryDateDisplay/CardExpiryDateDisplay.tsx
