+++ src/feature/CardRegisterPage/components/CardPreview/CardNumberDisplay/CardNumberDisplay.tsx
