+++ src/feature/CardRegisterPage/components/CardPreview/CardPreview.tsx
