+++ src/feature/CardRegisterPage/components/CardRegisterForm/shared/Label/Label.stories.tsx
