+++ src/feature/CardRegisterPage/utils/placeHolder.ts
