+++ src/feature/CardRegisterPage/validators/ErrorInformationType.ts
