+++ src/feature/CardRegisterPage/validators/cardCompany.ts
