+++ src/feature/CardRegisterPage/validators/cardNumber.ts
