+++ src/feature/CardRegisterPage/validators/password.ts
